% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf and pacf.R
\name{acf_plot}
\alias{acf_plot}
\title{acf and pacf plot.
This is the acf and pacf plot function.}
\usage{
acf_plot()
}
\arguments{
\item{x}{time series data}
}
\value{
the graphs of acf and pacf
}
\description{
acf and pacf plot.
This is the acf and pacf plot function.
}
\examples{

}
