% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosales data.R
\docType{data}
\name{autosales_data}
\alias{autosales_data}
\title{Autosales data set.}
\format{
An object of class \code{data.frame} with 341 rows and 5 columns.
}
\usage{
autosales_data
}
\arguments{
\item{x}{both cpi and auto sales could be used as x or y}

\item{y}{both cpi and auto sales could be used as x or y}
}
\description{
This is the autosales and cpi monthly data from 1970 to 1998.
}
\section{date, cpi, autosales}{
NA
}

\keyword{datasets}
