% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power Calculation.R
\name{pow}
\alias{pow}
\title{Power Calculation}
\usage{
pow(x, y)
}
\arguments{
\item{x}{a number in calculation}

\item{y}{the number raised to x}
}
\description{
function to print x raised to the power y
}
