% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum of squares.R
\name{squre_function}
\alias{squre_function}
\title{sum of squares}
\usage{
squre_function(x, y)
}
\arguments{
\item{two}{numbers-a,b}
}
\value{
square of (a+b)
}
\description{
This function calculate the sum of squares between a and b
}
