% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_var_sd.R
\name{wt_mean}
\alias{wt_mean}
\title{Mean, Variance, and standard deviation calculation.}
\usage{
wt_mean(x, w)
}
\arguments{
\item{input}{datagroup and weights for calculation}

\item{Mean, }{Variance, and standard deviation}
}
\description{
There are three functions calculate mean, var, and sd for dataset.
}
